\name{lpc-internal}
\title{Internal lpc functions}
\alias{print.lpcobj}
\alias{print.lpcfdrobj}
\alias{print.tfdrobj}
\description{Internal lpc functions}
%\usage{
%print.lpcobj(x,...)
%print.lpcfdrobj(x,...)
%print.tfdrobj(x,...)
%}
\author{Daniela Witten and Robert Tibshirani}
\details{These functions are not to be called by the user.}
\keyword{internal}




